airflow/requirements.txt
apache-airflow
kafka-python
requests